Replaced float data types to double in libobjrenderer and libpgmodeler_ui classes to avoid lost of precision mainly when handle with graphics.
Minor fix on range based loops using auto

+++ libobjrenderer/src/graphicalview.cpp
@@ -41,7 +41,7 @@ void GraphicalView::configureObject(void)
   RoundedRectItem *bodies[]={ body, ext_attribs_body };
   QString attribs[]={ ParsersAttributes::VIEW_BODY, ParsersAttributes::VIEW_EXT_BODY },
           tag_attribs[]={ ParsersAttributes::TABLE_BODY, ParsersAttributes::TABLE_EXT_BODY };
-	float width, type_width=0, px=0;
+	double width, type_width=0, px=0;
 	TableObjectView *col_item=nullptr;
 	QList<TableObjectView *> col_items;
 	TableObject *tab_obj=nullptr;
